Refactor expense entry and schedule management
Replaced `ExpenseEntryApiDelegateImpl.kt` with new expense entry components and API implementation. Introduced UI dialogs and tables for managing entries, refactored data models, and updated repositories for MongoDB integration.

+++ server/src/main/kotlin/com/jakubspiewak/sheep/report/ReportApiDelegateImpl.kt
@@ -1,7 +1,7 @@
 package com.jakubspiewak.sheep.report
 
-import com.jakubspiewak.sheep.expense.ExpenseScheduleDocument
-import com.jakubspiewak.sheep.expense.ExpenseScheduleRepository
+import com.jakubspiewak.sheep.expense.schedule.ExpenseScheduleDocument
+import com.jakubspiewak.sheep.expense.schedule.ExpenseScheduleRepository
 import com.jakubspiewak.sheep.generated.api.ReportApiDelegate
 import com.jakubspiewak.sheep.generated.model.*
 import com.jakubspiewak.sheep.income.IncomeBlueprintDocument
@@ -42,9 +42,9 @@ class ReportApiDelegateImpl(
         val totalIncomes = incomes.sumOf { it.getMonthlyAvgAmount() }
 
         val response = FinancialSummaryResponse(
-            totalIncome = totalIncomes.toFloat(),
-            totalExpenses = totalExpenses.toFloat(),
-            netBalance = totalIncomes.toFloat() - totalExpenses.toFloat()
+            totalIncome = totalIncomes,
+            totalExpenses = totalExpenses,
+            netBalance = totalIncomes - totalExpenses
         )
 
         return ResponseEntity.ok(response)
@@ -72,10 +72,10 @@ private fun ExpenseScheduleDocument.getMonthlyAvgAmount(): Double {
 
 private fun EstimatedAmount.getAvgValue(): Double {
     return when (this) {
-        is FixedAmount -> amount.toDouble()
+        is FixedAmount -> amount
         is RangeAmount -> (min + max) / 2.0
-        is VarianceAmount -> amount.toDouble()
-        is VariancePercent -> amount.toDouble()
+        is VarianceAmount -> amount
+        is VariancePercent -> amount
         else -> throw IllegalArgumentException("Unknown EstimatedAmount type")
     }
 }
